% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ChangeEntrezFromName}
\alias{ChangeEntrezFromName}
\title{change symbol to entrez}
\usage{
ChangeEntrezFromName(GeneCount)
}
\arguments{
\item{GeneCount}{}
}
\value{
gene entrez profile
}
\description{
change symbol to entrez
}
