% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CMSFFPEClassifier}
\alias{CMSFFPEClassifier}
\title{main function}
\usage{
CMSFFPEClassifier(Expr = NULL, minPosterior = 0.5, log2transfrom = FALSE)
}
\arguments{
\item{log2transfrom}{}
}
\value{
probability
}
\description{
main function
}
