% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PredictedInMultiRF}
\alias{PredictedInMultiRF}
\title{prediction}
\usage{
PredictedInMultiRF(model, texp, phenotype = NULL, minPosterior)
}
\arguments{
\item{minPosterior}{}
}
\value{
probability
}
\description{
prediction
}
